fix: 수익률 수정 및 출력 문구 추가

+++ src/main/java/lotto/domain/LottoResult.java
@@ -15,7 +15,7 @@ public class LottoResult {
     }
 
     private void calculateProfit(int lottoSize) {
-        totalProfit = totalReward / (lottoSize * LOTTO_PRICE);
+        totalProfit = (double) totalReward / (lottoSize * LOTTO_PRICE);
     }
 
     private void calculateTotalReward() {
